requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==42.0.8
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.96.1
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-auth==2.30.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.10.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kubernetes==29.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.79
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.5
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime-gpu==1.18.0
openai==1.34.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.16
pydantic_core==2.18.4
pydeck==0.8.0b4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.2
starlette==0.27.0
streamlit==1.35.0
sympy==1.12.1
tenacity==8.4.1
tiktoken==0.5.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-utils==0.1.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
uvicorn==0.30.1
validators==0.28.3
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
